refactor(b43): split the exam runner; state the MCC bound in sigma
Both remaining review items.

MODULE SIZE (codex P2). run.mjs was 523 lines against AGENTS.md's ~400. Split
into five focused modules -- stats, submission-run, diagnostics, brute-force,
salts -- leaving the runner at 377 and every module under the limit. Verified
as a pure code move: the exam was re-run and its scorecard compared field by
field against the committed one, 0 substantive differences. The only fields
that move are the wall-clock probe and runtimeSeconds.

That check caught its own vacuity once: the first comparison reported 0 diffs
while the run had actually died on a missing import, so it was comparing the
file to itself. The exit code is what caught it; the diff alone would have
read as success.

MCC THRESHOLD (CodeRabbit). `attackRetainsNoInformation` ANDed a bare
`|mcc| <= 0.05` with `|z| <= 2`, and the two disagreed: at n=1000 an absolute
0.05 is 1.58 sigma, so the arbitrary constant was the binding one and exam-C
passed by 3% of it. A correlation threshold that does not scale with n is a
coincidence, not a threshold -- the same figure would be damning at n=100 and
meaningless at n=100000. The bound is now 2 sigma = 2/sqrt(n), derived from the
model count, which is the z clause in MCC units so the two cannot drift apart.

A third clause was ADDED, not relaxed: the observed correlation must also lie
inside the range the nulls actually produced. exam-C motivates it -- its
0.048437 is not even the largest the nulls reach (0.072645), which is a stronger
statement than any threshold. Verdicts are unchanged; mutation-checked by
tightening the bound 10x, which flips it false.

The scorecard now records mccTwoSigmaBound, mccWorstArmMarginSigma and
aggregateOutlierArms, so a reader sees WHY exam-C sits at 24-of-24 on the
aggregate next to a PASS rather than finding an unexplained outlier.

Also corrected the same impossibility overclaim in REPORT.md that was fixed in
BENCHMARK.md: the clause is reachable by sabotage, not unreachable.
Re-blessed the one wall-clock figure the re-run moved (657.2 core-hours).

Gate 0 findings. 21 tests. Determinism 20/20.
